+++ lib/routing_path.dart
@@ -2,7 +2,7 @@ library routing_path;
 
 export 'src/errors.dart';
 export 'src/navigation_route_handler.dart';
-export 'src/path_route_handler.dart';
 export 'src/path_router.dart';
+export 'src/pattern_route_handler.dart';
 export 'src/route_arguments.dart';
 export 'src/route_handler.dart';